+++ Dsp/widget/wgt_core/widget_mng.c
@@ -1110,10 +1110,10 @@ static int8_t WidgetUI_GetCoord(const WidgetUI_Item_TypeDef *item, WidgetUI_GetG
 static WidgetUI_Item_TypeDef *WidgetUI_InsertSequence_Callback(const WidgetUI_Item_TypeDef *item_prv, const WidgetUI_Item_TypeDef *item_nxt)
 {
     /* this 4 var down below can be optmized*/
-    int8_t item_prv_y = 0;
-    int8_t item_nxt_y = 0;
-    int8_t item_prv_x = 0;
-    int8_t item_nxt_x = 0;
+    int16_t item_prv_y = 0;
+    int16_t item_nxt_y = 0;
+    int16_t item_prv_x = 0;
+    int16_t item_nxt_x = 0;
 
     if ((item_prv == NULL) && (item_nxt == NULL))
         return NULL;
